Deployment/requirements.txt
torch
torchvision
segmentation-models-pytorch
timm
gradio
pillow
numpy
grad-cam